+++ app/src/store/slices/userSlice.ts
@@ -146,13 +146,13 @@ export const selectDisplayName = createSelector(userState, (state) => state.disp
 export const selectAvatarUrl = createSelector(userState, (state) => state.avatarUrl);
 export const selectIsPresenter = createSelector(userState, (state) => state.isPresenter);
 export const selectUserProtocolAccess = createSelector(userState, (state) => state.protocolAccess);
-export const selectIsGuest = createSelector(userState, (state) => state.role === Role.Guest);
-export const selectIsModerator = createSelector(userState, (state) => state.role === Role.Moderator);
 export const selectIsLoggedIn = createSelector(userState, (state) => state.loggedIdToken !== undefined);
 export const selectIsAuthenticated = createSelector(
   userState,
   (state) => state.loggedIdToken !== undefined || state.role === Role.Guest
 );
+export const selectIsModerator = createSelector(userState, (state) => state.role === Role.Moderator);
+export const selectIsGuest = createSelector(userState, (state) => state.role === Role.Guest);
 
 export const selectUserAsPartialParticipant = createSelector(
   userState,